reminder cog type checking fix

+++ cogs/reminder.py
@@ -134,7 +134,7 @@ class Reminder(commands.Cog):
       self._task.cancel()
       self._task = self.bot.loop.create_task(self.dispatch_timers())
 
-  async def short_timer_optimisation(self, seconds: int, timer: Timer) -> None:
+  async def short_timer_optimisation(self, seconds: float, timer: Timer) -> None:
     await asyncio.sleep(seconds)
     event_name = f'{timer.event}_timer_complete'
     self.bot.dispatch(event_name, timer)
